% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{detect_format}
\alias{detect_format}
\title{Identify the storage format of a single file}
\usage{
detect_format(path)
}
\arguments{
\item{path}{File path.}
}
\value{
One of "mtx", "h5ad", "loom", "h5", "rds", "tar", "zip", "dense",
or "unknown".
}
\description{
Identify the storage format of a single file
}
\keyword{internal}
