% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{R2scGEO-package}
\alias{R2scGEO}
\alias{R2scGEO-package}
\title{R2scGEO: Query and Load Public scRNA-seq Datasets from GEO}
\description{
A lightweight toolkit to search NCBI GEO for single-cell RNA-seq series, inspect their metadata, list and download supplementary files, and load 10x-style sparse matrices into Seurat or SingleCellExperiment objects. Downloads are cached on disk to avoid repeated queries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ryanreis333/R2scGEO}
  \item Report bugs at \url{https://github.com/ryanreis333/R2scGEO/issues}
}

}
\author{
\strong{Maintainer}: Ryan Reis \email{ryanreis333@gmail.com}

}
\keyword{internal}
