% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{scgeo_cache_dir}
\alias{scgeo_cache_dir}
\title{Location of the on-disk download cache}
\usage{
scgeo_cache_dir()
}
\value{
Path to the cache directory (created on first use).
}
\description{
Location of the on-disk download cache
}
