% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{scgeo_clear_cache}
\alias{scgeo_clear_cache}
\title{Clear the scgeo download cache}
\usage{
scgeo_clear_cache(accession = NULL)
}
\arguments{
\item{accession}{Optional accession to clear; if \code{NULL}, clears everything.}
}
\value{
Invisibly, \code{TRUE}.
}
\description{
Clear the scgeo download cache
}
