% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{scgeo_suppl_files}
\alias{scgeo_suppl_files}
\title{List supplementary files available for a GEO accession}
\usage{
scgeo_suppl_files(accession)
}
\arguments{
\item{accession}{A GSE/GSM accession, e.g. "GSE164897".}
}
\value{
A tibble with \code{file} and \code{url} columns.
}
\description{
Parses the Apache directory listing of the accession's GEO FTP \verb{suppl/}
folder.
}
